move tests from tdf3 into test folder
