+++ lib/.mocharc.json
@@ -1,5 +1,5 @@
 {
   "extension": ["ts"],
-  "spec": ["tdf3/test/**/*.spec.+(js|ts)"],
-  "require": "tdf3/test/babel-register.js"
+  "spec": ["test/mocha/**/*.spec.+(js|ts)"],
+  "require": "test/mocha/babel-register.js"
 }
