+++ lib/test/mocha/client.spec.js
@@ -2,8 +2,8 @@ import { assert } from 'chai';
 import { sync as rimrafSync } from 'rimraf';
 import { mkdirSync, readFileSync } from 'fs';
 
-import { Client as TDF } from '../src';
-import { NodeTdfStream } from '../src/client/NodeTdfStream';
+import { Client as TDF } from '../../tdf3/src';
+import { NodeTdfStream } from '../../tdf3/src/client/NodeTdfStream';
 const defaultConfig = {
   clientId: 'id',
   kasEndpoint: 'kas',
