+++ lib/test/mocha/perf.spec.js
@@ -11,7 +11,7 @@ import JSZip from 'jszip';
 import { pipeline } from 'stream';
 import { createSandbox } from 'sinon';
 
-import { Client, TDF } from '../src';
+import { Client, TDF } from '../../tdf3/src/index';
 import getMocks from '../mocks';
 
 const Mocks = getMocks();
