+++ lib/test/mocha/setup.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-undef */
 import { webcrypto } from 'crypto';
 import { ReadableStream } from 'stream/web';
-import { registerModuleType } from '../src/client/tdf-stream';
-import { NodeTdfStream } from '../src/client/NodeTdfStream';
+import { registerModuleType } from '../../tdf3/src/client/tdf-stream';
+import { NodeTdfStream } from '../../tdf3/src/client/NodeTdfStream';
 
 if (typeof globalThis.ReadableStream === 'undefined') {
   globalThis.ReadableStream = ReadableStream;
