+++ lib/test/mocha/unit/builders.spec.js
@@ -1,5 +1,5 @@
 import { expect, assert } from 'chai';
-import { EncryptParamsBuilder } from '../../src/client/builders';
+import { EncryptParamsBuilder } from '../../../tdf3/src/client/builders';
 
 describe('EncyptParamsBuilder', () => {
   describe('setAttributes', () => {
