+++ lib/test/mocha/unit/crypto/crypto-utils.spec.js
@@ -4,7 +4,7 @@ import {
   formatAsPem,
   isValidAsymmetricKeySize,
   removePemFormatting,
-} from '../../../src/crypto/crypto-utils';
+} from '../../../../tdf3/src/crypto/crypto-utils';
 
 describe('crypto-utils', () => {
   it('should remove pem formatting', () => {
