+++ lib/test/mocha/unit/errors.spec.js
@@ -9,7 +9,7 @@ import {
   TdfDecryptError,
   TdfError,
   TdfPayloadExtractionError,
-} from '../../src/errors';
+} from '../../../tdf3/src/errors';
 
 describe('Errors', () => {
   const errorClasses = {
