+++ lib/test/mocha/unit/mock-stream.spec.js
@@ -1,5 +1,5 @@
 import { assert } from 'chai';
-import { MockStream } from '../../src/utils/mock-stream';
+import { MockStream } from '../../../tdf3/src/utils/mock-stream';
 
 describe('MockStream', () => {
   it('constructs empty with no arguments', () => {
