+++ lib/test/mocha/unit/tdf.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { TDF } from '../../src';
+import { TDF } from '../../../tdf3/src';
 
 describe('TDF', () => {
   it('constructs', () => {
