+++ lib/test/mocha/unit/templates/escaper.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { escHtml, escJavaScript } from '../../../src/templates/index';
+import { escHtml, escJavaScript } from '../../../../tdf3/src/templates/index';
 
 describe('Validate html escaping', () => {
   it('esc no op', () => {
