+++ lib/test/mocks/client/attr_policy_config.json
