+++ lib/test/mocks/client/default_manifest.json
