+++ lib/test/mocks/client/default_policy.json
