+++ lib/test/mocks/client/public_key_lookup.json
