+++ lib/test/mocks/kas/rewrap/response-simple.json
